Add kUpdateVdexNeeded for DexOptNeeded.
And as consequence, add --update-vdex to dex2oat (which currently
has no effect).

ART part of this change.

bug: 30937355
Test: oat_file_assistant_test

+++ runtime/oat_file_assistant.cc
@@ -178,8 +178,18 @@ OatFileAssistant::GetDexOptNeeded(CompilerFilter::Filter target, bool profile_ch
     }
   }
 
-  // We can only run dex2oat if there are original dex files.
-  return HasOriginalDexFiles() ? kDex2OatNeeded : kNoDexOptNeeded;
+  if (!HasOriginalDexFiles()) {
+    // We can only run dex2oat if there are original dex files.
+    return kNoDexOptNeeded;
+  }
+
+  if (OatFileExists() || OdexFileExists()) {
+    if (!IsUpToDate()) {
+      return kUpdateVdexNeeded;
+    }
+  }
+
+  return kDex2OatNeeded;
 }
 
 // Figure out the currently specified compile filter option in the runtime.
@@ -217,9 +227,10 @@ OatFileAssistant::MakeUpToDate(bool profile_changed, std::string* error_msg) {
 
   switch (GetDexOptNeeded(target, profile_changed)) {
     case kNoDexOptNeeded: return kUpdateSucceeded;
-    case kDex2OatNeeded: return GenerateOatFile(error_msg);
+    case kDex2OatNeeded: return GenerateOatFile(/* update_vdex */ false, error_msg);
     case kPatchOatNeeded: return RelocateOatFile(odex_.Filename(), error_msg);
     case kSelfPatchOatNeeded: return RelocateOatFile(oat_.Filename(), error_msg);
+    case kUpdateVdexNeeded: return GenerateOatFile(/* update_vdex */ true, error_msg);
   }
   UNREACHABLE();
 }
@@ -538,7 +549,7 @@ OatFileAssistant::RelocateOatFile(const std::string* input_file, std::string* er
 }
 
 OatFileAssistant::ResultOfAttemptToUpdate
-OatFileAssistant::GenerateOatFile(std::string* error_msg) {
+OatFileAssistant::GenerateOatFile(bool update_vdex, std::string* error_msg) {
   CHECK(error_msg != nullptr);
 
   Runtime* runtime = Runtime::Current();
@@ -564,19 +575,31 @@ OatFileAssistant::GenerateOatFile(std::string* error_msg) {
     return kUpdateNotAttempted;
   }
 
-  std::unique_ptr<File> vdex_file(OS::CreateEmptyFile(vdex_file_name.c_str()));
-  if (vdex_file.get() == nullptr) {
-    *error_msg = "Generation of oat file " + oat_file_name
-      + " not attempted because the vdex file " + vdex_file_name
-      + " could not be opened.";
-    return kUpdateNotAttempted;
-  }
+  std::unique_ptr<File> vdex_file(nullptr);
+  update_vdex = update_vdex && OS::FileExists(vdex_file_name.c_str());
+  if (update_vdex) {
+    vdex_file.reset(OS::OpenFileReadWrite(vdex_file_name.c_str()));
+    if (vdex_file.get() == nullptr) {
+      *error_msg = "Generation of oat file " + oat_file_name
+        + " not attempted because the vdex file " + vdex_file_name
+        + " could not be opened.";
+      return kUpdateNotAttempted;
+    }
+  } else {
+    vdex_file.reset(OS::CreateEmptyFile(vdex_file_name.c_str()));
+    if (vdex_file.get() == nullptr) {
+      *error_msg = "Generation of oat file " + oat_file_name
+        + " not attempted because the vdex file " + vdex_file_name
+        + " could not be opened.";
+      return kUpdateNotAttempted;
+    }
 
-  if (fchmod(vdex_file->Fd(), 0644) != 0) {
-    *error_msg = "Generation of oat file " + oat_file_name
-      + " not attempted because the vdex file " + vdex_file_name
-      + " could not be made world readable.";
-    return kUpdateNotAttempted;
+    if (fchmod(vdex_file->Fd(), 0644) != 0) {
+      *error_msg = "Generation of oat file " + oat_file_name
+        + " not attempted because the vdex file " + vdex_file_name
+        + " could not be made world readable.";
+      return kUpdateNotAttempted;
+    }
   }
 
   std::unique_ptr<File> oat_file(OS::CreateEmptyFile(oat_file_name.c_str()));
@@ -598,12 +621,18 @@ OatFileAssistant::GenerateOatFile(std::string* error_msg) {
   args.push_back("--vdex-fd=" + std::to_string(vdex_file->Fd()));
   args.push_back("--oat-fd=" + std::to_string(oat_file->Fd()));
   args.push_back("--oat-location=" + oat_file_name);
+  if (update_vdex) {
+    args.push_back("--update-vdex");
+  }
 
   if (!Dex2Oat(args, error_msg)) {
     // Manually delete the oat and vdex files. This ensures there is no garbage
-    // left over if the process unexpectedly died.
-    vdex_file->Erase();
-    unlink(vdex_file_name.c_str());
+    // left over if the process unexpectedly died. If `update_vdex` is true, we
+    // do not delete the vdex file as it was existing prior to this invocation.
+    if (!update_vdex) {
+      vdex_file->Erase();
+      unlink(vdex_file_name.c_str());
+    }
     oat_file->Erase();
     unlink(oat_file_name.c_str());
     return kUpdateFailed;
